Break downloading data into separate module
This prevents `requests` form being imported unless it is needed.
`requests` hangs on the b512 node on Anvil/Blues.
