+++ download_natural_earth_110m.py
@@ -4,7 +4,7 @@ from __future__ import absolute_import, division, print_function, \
     unicode_literals
 
 
-from mpas_analysis.__main__ import download_natural_earth_110m
+from mpas_analysis.download_data import download_natural_earth_110m
 
 
 if __name__ == "__main__":
