+++ mpas_analysis/__main__.py
@@ -34,7 +34,6 @@ import progressbar
 import logging
 import xarray
 import time
-import cartopy.io.shapereader as shpreader
 
 from mpas_analysis.shared.analysis_task import AnalysisFormatter
 
@@ -57,8 +56,6 @@ from mpas_analysis import sea_ice
 from mpas_analysis.shared.climatology import MpasClimatologyTask, \
     RefYearMpasClimatologyTask
 from mpas_analysis.shared.time_series import MpasTimeSeriesTask
-
-from mpas_analysis.shared.io.download import download_files
 
 
 def update_time_bounds_in_config(config):  # {{{
@@ -857,47 +854,6 @@ def main():
 
     if not args.setup_only:
         generate_html(config, analyses, controlConfig)
-
-
-def download_analysis_data():
-    """
-    Entry point for downloading the input data set from public repository for
-    MPAS-Analysis to work. The input data set includes: pre-processed
-    observations data, MPAS mapping files and MPAS regional mask files
-    (which are used for the MOC computation), for a subset of MPAS meshes.
-    """
-
-    parser = argparse.ArgumentParser(
-        description=__doc__, formatter_class=argparse.RawTextHelpFormatter)
-    parser.add_argument("-o", "--outDir", dest="outDir", required=True,
-                        help="Directory where MPAS-Analysis input data will"
-                             "be downloaded")
-    parser.add_argument("-d", "--dataset", dest="dataset", default='analysis',
-                        help="Directory where MPAS-Analysis input data will"
-                             "be downloaded")
-    args = parser.parse_args()
-
-    try:
-        os.makedirs(args.outDir)
-    except OSError:
-        pass
-
-    urlBase = 'https://web.lcrc.anl.gov/public/e3sm/diagnostics'
-    analysisFileList = pkg_resources.resource_string(
-        'mpas_analysis',
-        'obs/{}_input_files'.format(args.dataset)).decode('utf-8')
-
-    # remove any empty strings from the list
-    analysisFileList = list(filter(None, analysisFileList.split('\n')))
-    download_files(analysisFileList, urlBase, args.outDir, verify=True)
-
-
-def download_natural_earth_110m():
-    for name in ['ocean', 'coastline', 'land']:
-        shpfilename = shpreader.natural_earth(resolution='110m',
-                                              category='physical',
-                                              name=name)
-        reader = shpreader.Reader(shpfilename)
 
 
 if __name__ == "__main__":
